refactor: Rank와 Board의 패키지를 board의 하위로 이동
