+++ src/test/java/chess/board/BoardTest.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.board;
 
 import chess.pieces.Piece;
 import org.junit.jupiter.api.DisplayName;
